pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "siggate-gt"
dynamic = ["version"]
description = "SigGate-GT: Element-wise sigmoid gating on graph transformer attention to address over-smoothing"
readme = "README.md"
requires-python = ">=3.10,<3.12"
authors = [
    {name = "Anonymous", email = "anonymous@example.com"},
]
keywords = [
    "graph transformers",
    "gated attention",
    "over-smoothing",
    "molecular property prediction",
    "deep learning",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]

dependencies = [
    "torch>=2.1.0",
    "torch-geometric>=2.4.0",
    "numpy>=1.24.0,<2.0.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    "ogb>=1.3.6",
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "wandb>=0.16.0",
    "tqdm>=4.66.0",
    "rich>=13.7.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.5.0",
    "pytest-timeout>=2.2.0",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
    "ipython>=8.19.0",
    "ipdb>=0.13.13",
]
docs = [
    "mkdocs>=1.5.3",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
]
all = [
    "siggate-gt[dev,docs]",
]

[project.urls]
Repository = "https://anonymous.4open.science/r/SigGate-GT"
Issues = "https://anonymous.4open.science/r/SigGate-GT/issues"

[tool.setuptools.packages.find]
where = ["src"]
include = ["siggate_gt*"]

[tool.setuptools.dynamic]
version = {attr = "siggate_gt.__version__.__version__"}

[tool.setuptools.package-data]
"siggate_gt" = ["py.typed", "*.yaml", "*.json"]

# ─────────────────────────────────────────────────────────────────────────────
# RUFF
# ─────────────────────────────────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]
exclude = [
    ".git", ".mypy_cache", ".ruff_cache", ".venv", "venv",
    "__pycache__", "build", "dist",
]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "B", "C4", "SIM", "RUF"]
ignore = ["E501", "PLR0913", "PLR2004"]

[tool.ruff.lint.isort]
known-first-party = ["siggate_gt"]
force-single-line = false
lines-after-imports = 2

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["PLR2004", "S101"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

# ─────────────────────────────────────────────────────────────────────────────
# MYPY
# ─────────────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
check_untyped_defs = true
show_error_codes = true
pretty = true

[[tool.mypy.overrides]]
module = [
    "torch.*", "torch_geometric.*", "ogb.*",
    "sklearn.*", "scipy.*", "numpy.*",
    "wandb.*", "hydra.*", "omegaconf.*",
    "tqdm.*", "rich.*",
]
ignore_missing_imports = true

# ─────────────────────────────────────────────────────────────────────────────
# PYTEST
# ─────────────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "-ra",
    "--tb=short",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "gpu: marks tests requiring GPU (deselect with '-m \"not gpu\"')",
    "integration: marks integration tests",
    "unit: marks unit tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]
timeout = 120

# ─────────────────────────────────────────────────────────────────────────────
# COVERAGE
# ─────────────────────────────────────────────────────────────────────────────
[tool.coverage.run]
source = ["src"]
branch = true
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "@abstractmethod",
]
fail_under = 60
show_missing = true
